Moved tests for pager.py to new test_pager.py file from test_command_toolbar.py
Also:
- Moved common code used by test_command_toolbar.py and test_pager.py to conftest.py

Since the pager is effectively its own component, it feels cleaner to have the tests in their own file.
